更新原作者内容201008

+++ qt/aqt/profiles.py
@@ -11,19 +11,22 @@ import locale
 import pickle
 import random
 import shutil
-from typing import Any, Dict
+import traceback
+from typing import Any, Dict, List, Optional
 
 from send2trash import send2trash
 
 import anki.lang
 import aqt.forms
 import aqt.sound
+from anki import Collection
 from anki.db import DB
-from anki.lang import _
+from anki.lang import _, without_unicode_isolation
+from anki.rsbackend import SyncAuth
 from anki.utils import intTime, isMac, isWin
 from aqt import appHelpSite
 from aqt.qt import *
-from aqt.utils import locale_dir, showWarning
+from aqt.utils import TR, locale_dir, showWarning, tr
 
 metaConf = dict(
     ver=0,
@@ -34,7 +37,6 @@ metaConf = dict(
     suppressUpdate=False,
     firstRun=True,
     defaultLang=None,
-    disabledAddons=[],
 )
 
 profileConf: Dict[str, Any] = dict(
@@ -67,14 +69,22 @@ class LoadMetaResult:
     loadError: bool
 
 
+class AnkiRestart(SystemExit):
+    def __init__(self, *args, **kwargs):
+        self.exitcode = kwargs.pop("exitcode", 0)
+        super().__init__(*args, **kwargs)  # type: ignore
+
+
 class ProfileManager:
     def __init__(self, base=None):
+        ## Settings which should be forgotten each Anki restart
+        self.session = {}
         self.name = None
         self.db = None
+        self.profile: Optional[Dict] = None
         # instantiate base folder
+        self.base: str
         self._setBaseFolder(base)
-
-        aqt.sound.setMpvConfigBase(self.base)
 
     def setupMeta(self) -> LoadMetaResult:
         # load metadata
@@ -83,26 +93,26 @@ class ProfileManager:
         return res
 
     # profile load on startup
-    def openProfile(self, profile):
+    def openProfile(self, profile) -> None:
         if profile:
             if profile not in self.profiles():
                 QMessageBox.critical(None, "Error", "Requested profile does not exist.")
                 sys.exit(1)
             try:
                 self.load(profile)
-            except TypeError:
-                raise Exception("Provided profile does not exist.")
+            except TypeError as exc:
+                raise Exception("Provided profile does not exist.") from exc
 
     # Base creation
     ######################################################################
 
-    def ensureBaseExists(self):
+    def ensureBaseExists(self) -> None:
         self._ensureExists(self.base)
 
     # Folder migration
     ######################################################################
 
-    def _oldFolderLocation(self):
+    def _oldFolderLocation(self) -> str:
         if isMac:
             return os.path.expanduser("~/Documents/Anki")
         elif isWin:
@@ -116,16 +126,83 @@ class ProfileManager:
             return os.path.expanduser("~/Documents/Anki")
 
     def maybeMigrateFolder(self):
+        newBase = self.base
         oldBase = self._oldFolderLocation()
 
         if oldBase and not os.path.exists(self.base) and os.path.isdir(oldBase):
+            try:
+                # if anything goes wrong with UI, reset to the old behavior of always migrating
+                self._tryToMigrateFolder(oldBase)
+            except AnkiRestart:
+                raise
+            except:
+                print("migration failed")
+                self.base = newBase
+                shutil.move(oldBase, self.base)
+
+    def _tryToMigrateFolder(self, oldBase):
+        from PyQt5 import QtGui, QtWidgets
+
+        app = QtWidgets.QApplication([])
+        icon = QtGui.QIcon()
+        icon.addPixmap(
+            QtGui.QPixmap(":/icons/anki.png"),
+            QtGui.QIcon.Normal,
+            QtGui.QIcon.Off,
+        )
+        window_title = "Data Folder Migration"
+        migration_directories = f"\n\n    {oldBase}\n\nto\n\n    {self.base}"
+
+        confirmation = QMessageBox()
+        confirmation.setIcon(QMessageBox.Warning)
+        confirmation.setWindowIcon(icon)
+        confirmation.setStandardButtons(QMessageBox.Ok | QMessageBox.Cancel)  # type: ignore
+        confirmation.setWindowTitle(window_title)
+        confirmation.setText(
+            "Anki needs to move its data folder from Documents/Anki to a new location. Proceed?"
+        )
+        retval = confirmation.exec()
+
+        if retval == QMessageBox.Ok:
+            progress = QMessageBox()
+            progress.setIcon(QMessageBox.Information)
+            progress.setStandardButtons(QMessageBox.NoButton)
+            progress.setWindowIcon(icon)
+            progress.setWindowTitle(window_title)
+            progress.setText("Please wait...")
+            progress.show()
+            app.processEvents()  # type: ignore
+
             shutil.move(oldBase, self.base)
+            progress.hide()
+
+            completion = QMessageBox()
+            completion.setIcon(QMessageBox.Information)
+            completion.setStandardButtons(QMessageBox.Ok)
+            completion.setWindowIcon(icon)
+            completion.setWindowTitle(window_title)
+            completion.setText("Migration complete. Please start Anki again.")
+            completion.show()
+            completion.exec()
+        else:
+            diag = QMessageBox()
+            diag.setIcon(QMessageBox.Warning)
+            diag.setWindowIcon(icon)
+            diag.setStandardButtons(QMessageBox.Ok)
+            diag.setWindowTitle(window_title)
+            diag.setText(
+                "Migration aborted. If you would like to keep the old folder location, please "
+                "see the Startup Options section of the manual. Anki will now quit."
+            )
+            diag.exec()
+
+        raise AnkiRestart(exitcode=0)
 
     # Profile load/save
     ######################################################################
 
-    def profiles(self):
-        def names():
+    def profiles(self) -> List:
+        def names() -> List:
             return self.db.list("select name from profiles where name != '_global'")
 
         n = names()
@@ -135,12 +212,12 @@ class ProfileManager:
 
         return n
 
-    def _unpickle(self, data):
+    def _unpickle(self, data) -> Any:
         class Unpickler(pickle.Unpickler):
-            def find_class(self, module, name):
+            def find_class(self, module: str, name: str) -> Any:
                 if module == "PyQt5.sip":
                     try:
-                        import PyQt5.sip  # type: ignore # pylint: disable=unused-import
+                        import PyQt5.sip  # pylint: disable=unused-import
                     except:
                         # use old sip location
                         module = "sip"
@@ -160,10 +237,10 @@ class ProfileManager:
         up = Unpickler(io.BytesIO(data), errors="ignore")
         return up.load()
 
-    def _pickle(self, obj):
+    def _pickle(self, obj) -> Any:
         return pickle.dumps(obj, protocol=0)
 
-    def load(self, name):
+    def load(self, name) -> bool:
         assert name != "_global"
         data = self.db.scalar(
             "select cast(data as blob) from profiles where name = ?", name
@@ -181,38 +258,38 @@ Anki could not read your profile data. Window sizes and your sync login \
 details have been forgotten."""
                 ),
             )
-
+            traceback.print_stack()
             print("resetting corrupt profile")
             self.profile = profileConf.copy()
             self.save()
         return True
 
-    def save(self):
+    def save(self) -> None:
         sql = "update profiles set data = ? where name = ?"
         self.db.execute(sql, self._pickle(self.profile), self.name)
         self.db.execute(sql, self._pickle(self.meta), "_global")
         self.db.commit()
 
-    def create(self, name):
+    def create(self, name) -> None:
         prof = profileConf.copy()
         self.db.execute(
             "insert or ignore into profiles values (?, ?)", name, self._pickle(prof)
         )
         self.db.commit()
 
-    def remove(self, name):
+    def remove(self, name) -> None:
         p = self.profileFolder()
         if os.path.exists(p):
             send2trash(p)
         self.db.execute("delete from profiles where name = ?", name)
         self.db.commit()
 
-    def trashCollection(self):
+    def trashCollection(self) -> None:
         p = self.collectionPath()
         if os.path.exists(p):
             send2trash(p)
 
-    def rename(self, name):
+    def rename(self, name) -> None:
         oldName = self.name
         oldFolder = self.profileFolder()
         self.name = name
@@ -263,30 +340,52 @@ and no other programs are accessing your profile folders, then try again."""
     # Folder handling
     ######################################################################
 
-    def profileFolder(self, create=True):
+    def profileFolder(self, create=True) -> str:
         path = os.path.join(self.base, self.name)
         if create:
             self._ensureExists(path)
         return path
 
-    def addonFolder(self):
+    def addonFolder(self) -> str:
         return self._ensureExists(os.path.join(self.base, "addons21"))
 
-    def backupFolder(self):
+    def backupFolder(self) -> str:
         return self._ensureExists(os.path.join(self.profileFolder(), "backups"))
 
-    def collectionPath(self):
+    def collectionPath(self) -> str:
         return os.path.join(self.profileFolder(), "collection.anki2")
+
+    # Downgrade
+    ######################################################################
+
+    def downgrade(self, profiles=List[str]) -> List[str]:
+        "Downgrade all profiles. Return a list of profiles that couldn't be opened."
+        problem_profiles = []
+        for name in profiles:
+            path = os.path.join(self.base, name, "collection.anki2")
+            if not os.path.exists(path):
+                continue
+            with DB(path) as db:
+                if db.scalar("select ver from col") == 11:
+                    # nothing to do
+                    continue
+            try:
+                c = Collection(path)
+                c.close(save=False, downgrade=True)
+            except Exception as e:
+                print(e)
+                problem_profiles.append(name)
+        return problem_profiles
 
     # Helpers
     ######################################################################
 
-    def _ensureExists(self, path):
+    def _ensureExists(self, path: str) -> str:
         if not os.path.exists(path):
             os.makedirs(path)
         return path
 
-    def _setBaseFolder(self, cmdlineBase):
+    def _setBaseFolder(self, cmdlineBase: None) -> None:
         if cmdlineBase:
             self.base = os.path.abspath(cmdlineBase)
         elif os.environ.get("ANKI_BASE"):
@@ -296,7 +395,7 @@ and no other programs are accessing your profile folders, then try again."""
             self.maybeMigrateFolder()
         self.ensureBaseExists()
 
-    def _defaultBase(self):
+    def _defaultBase(self) -> str:
         if isWin:
             from aqt.winpaths import get_appdata
 
@@ -318,12 +417,12 @@ and no other programs are accessing your profile folders, then try again."""
 
         opath = os.path.join(self.base, "prefs.db")
         path = os.path.join(self.base, "prefs21.db")
-        if os.path.exists(opath) and not os.path.exists(path):
+        if not retrying and os.path.exists(opath) and not os.path.exists(path):
             shutil.copy(opath, path)
 
         result.firstTime = not os.path.exists(path)
 
-        def recover():
+        def recover() -> None:
             # if we can't load profile, start with a new one
             if self.db:
                 try:
@@ -348,6 +447,7 @@ create table if not exists profiles
                 "select cast(data as blob) from profiles where name = '_global'"
             )
         except:
+            traceback.print_stack()
             if result.loadError:
                 # already failed, prevent infinite loop
                 raise
@@ -361,8 +461,10 @@ create table if not exists profiles
                 self.meta = self._unpickle(data)
                 return result
             except:
+                traceback.print_stack()
                 print("resetting corrupt _global")
                 result.loadError = True
+                result.firstTime = True
 
         # if new or read failed, create a default global profile
         self.meta = metaConf.copy()
@@ -372,28 +474,22 @@ create table if not exists profiles
         )
         return result
 
-    def _ensureProfile(self):
+    def _ensureProfile(self) -> None:
         "Create a new profile if none exists."
         self.create(_("User 1"))
         p = os.path.join(self.base, "README.txt")
-        open(p, "w", encoding="utf8").write(
-            _(
-                """\
-This folder stores all of your Anki data in a single location,
-to make backups easy. To tell Anki to use a different location,
-please see:
-
-%s
-"""
+        with open(p, "w", encoding="utf8") as file:
+            file.write(
+                without_unicode_isolation(
+                    tr(TR.PROFILES_FOLDER_README, link=appHelpSite + "#startupopts")
+                )
             )
-            % (appHelpSite + "#startupopts")
-        )
 
     # Default language
     ######################################################################
     # On first run, allow the user to choose the default language
 
-    def setDefaultLang(self):
+    def setDefaultLang(self) -> None:
         # create dialog
         class NoCloseDiag(QDialog):
             def reject(self):
@@ -402,8 +498,8 @@ please see:
         d = self.langDiag = NoCloseDiag()
         f = self.langForm = aqt.forms.setlang.Ui_Dialog()
         f.setupUi(d)
-        d.accepted.connect(self._onLangSelected)
-        d.rejected.connect(lambda: True)
+        qconnect(d.accepted, self._onLangSelected)
+        qconnect(d.rejected, lambda: True)
         # default to the system language
         try:
             (lang, enc) = locale.getdefaultlocale()
@@ -426,33 +522,33 @@ please see:
         f.lang.setCurrentRow(idx)
         d.exec_()
 
-    def _onLangSelected(self):
+    def _onLangSelected(self) -> None:
         f = self.langForm
         obj = anki.lang.langs[f.lang.currentRow()]
         code = obj[1]
         name = obj[0]
         en = "Are you sure you wish to display Anki's interface in %s?"
         r = QMessageBox.question(
-            None, "Anki", en % name, QMessageBox.Yes | QMessageBox.No, QMessageBox.No
+            None, "Anki", en % name, QMessageBox.Yes | QMessageBox.No, QMessageBox.No  # type: ignore
         )
         if r != QMessageBox.Yes:
             return self.setDefaultLang()
         self.setLang(code)
 
-    def setLang(self, code):
+    def setLang(self, code) -> None:
         self.meta["defaultLang"] = code
         sql = "update profiles set data = ? where name = ?"
         self.db.execute(sql, self._pickle(self.meta), "_global")
         self.db.commit()
-        anki.lang.setLang(code, locale_dir(), local=False)
+        anki.lang.set_lang(code, locale_dir())
 
     # OpenGL
     ######################################################################
 
-    def _glPath(self):
+    def _glPath(self) -> str:
         return os.path.join(self.base, "gldriver")
 
-    def glMode(self):
+    def glMode(self) -> str:
         if isMac:
             return "auto"
 
@@ -460,7 +556,8 @@ please see:
         if not os.path.exists(path):
             return "software"
 
-        mode = open(path, "r").read().strip()
+        with open(path, "r") as file:
+            mode = file.read().strip()
 
         if mode == "angle" and isWin:
             return mode
@@ -468,10 +565,11 @@ please see:
             return mode
         return "auto"
 
-    def setGlMode(self, mode):
-        open(self._glPath(), "w").write(mode)
+    def setGlMode(self, mode) -> None:
+        with open(self._glPath(), "w") as file:
+            file.write(mode)
 
-    def nextGlMode(self):
+    def nextGlMode(self) -> None:
         mode = self.glMode()
         if mode == "software":
             self.setGlMode("auto")
@@ -483,11 +581,74 @@ please see:
         elif mode == "angle":
             self.setGlMode("software")
 
-    # Scale
+    # Shared options
     ######################################################################
 
     def uiScale(self) -> float:
-        return self.meta.get("uiScale", 1.0)
+        scale = self.meta.get("uiScale", 1.0)
+        return max(scale, 1)
 
     def setUiScale(self, scale: float) -> None:
         self.meta["uiScale"] = scale
+
+    def last_addon_update_check(self) -> int:
+        return self.meta.get("last_addon_update_check", 0)
+
+    def set_last_addon_update_check(self, secs) -> None:
+        self.meta["last_addon_update_check"] = secs
+
+    def night_mode(self) -> bool:
+        return self.meta.get("night_mode", False)
+
+    def set_night_mode(self, on: bool) -> None:
+        self.meta["night_mode"] = on
+
+    def dark_mode_widgets(self) -> bool:
+        return self.meta.get("dark_mode_widgets", False)
+
+    # Profile-specific
+    ######################################################################
+
+    def interrupt_audio(self) -> bool:
+        return self.profile.get("interrupt_audio", True)
+
+    def set_interrupt_audio(self, val: bool) -> None:
+        self.profile["interrupt_audio"] = val
+        aqt.sound.av_player.interrupt_current_audio = val
+
+    def set_sync_key(self, val: Optional[str]) -> None:
+        self.profile["syncKey"] = val
+
+    def set_sync_username(self, val: Optional[str]) -> None:
+        self.profile["syncUser"] = val
+
+    def set_host_number(self, val: Optional[int]) -> None:
+        self.profile["hostNum"] = val or 0
+
+    def media_syncing_enabled(self) -> bool:
+        return self.profile["syncMedia"]
+
+    def auto_syncing_enabled(self) -> bool:
+        return self.profile["autoSync"]
+
+    def sync_auth(self) -> Optional[SyncAuth]:
+        hkey = self.profile.get("syncKey")
+        if not hkey:
+            return None
+        return SyncAuth(hkey=hkey, host_number=self.profile.get("hostNum", 0))
+
+    def clear_sync_auth(self) -> None:
+        self.profile["syncKey"] = None
+        self.profile["syncUser"] = None
+        self.profile["hostNum"] = 0
+
+    def auto_sync_media_minutes(self) -> int:
+        return self.profile.get("autoSyncMediaMinutes", 15)
+
+    def set_auto_sync_media_minutes(self, val: int) -> None:
+        self.profile["autoSyncMediaMinutes"] = val
+
+    ######################################################################
+
+    def apply_profile_options(self) -> None:
+        aqt.sound.av_player.interrupt_current_audio = self.interrupt_audio()
